Add control messages to the stream
At any given time, there should really only be one active client. That
client should be either the one that is currently in the car, the one
that has most recently exited the car, or the one that has most recently
joined the session (in that order).

There are probably still some race conditions in the
LiveTelemetryPusherService, but this should be good enough for now.