Create Query language and Processors
Queries are used to query raw telemetry data. A query will be
transformed to a processor, which will process the data one point at
a time. This implements a very basic query language that supports only
variables and two functions, LAP_DELTA and LAP_AVERAGE.

LAP_DELTA calculates the difference between a value one lap ago and the
value now.

LAP_AVERAGE calculates the rolling average of a value for some number of
given laps.