Stream data live, using database triggers
Live data will be monitored using NOTIFY and LISTEN, which notify
receivers, rather than forcing receivers to poll. The NOTIFY messages
are sent automatically when rows are inserted into the table, and the
channels are named using the primary keys of the session (prefixed with
the initials of the table that they're coming from).

In Kotlin, we use BlockingQueues to send the notifications, which can be
used in the Store to wait for data and repeatedly query.