Create DECREASING_SUM for calculating usage over time
Fuel usage is the primary use case for this, summing the decreasing
intervals of FuelLevel will give the fuel usage over time. This replaces
the more simplistic LAP_DELTA method for computing the difference, as
that would take into account the refueling, ending up with a negative
number for a lap after refueling.