Replace React with Svelte
React doesn't work very well with live data. The hack I created to
handle the rapid pace of changes limited the code to 4 Hz. even that was
janky. Scrolling down on any page would cause the page to scroll back up
(probably fixable), but the underlying issue seems to be that rewriting
the whole DOM every quarter of a second is just not sustainable.

Svelte is far better suited for live data. Since it's compiled, it only
updates the element that changes, meaning that it is extremely smooth.
It's also much closer to simple JavaScript, meaning that I don't need to
find a wrapper library every time I want to do something.

I used Gemini to assist with converting the old frontend to Svelte.
Things aren't perfect, but it's roughly at feature parity. Subsequent
changes include things like shadcn-svelte and new features that were
relatively simple to integrate via Svelte.