Calculate laps remaining at the finish line
The laps remaining at the finish line is useful to know to determine if
fuel saving is worth it or not. It's computed by taking the current tank
level, subtracting out the margin and the expected fuel usage for the
rest of the lap, then by dividing that value by the average fuel usage.

This creates a new synthetic variable that is the amount of fuel used
for the current lap ("LapFuelUsed"), which we can subtract from the
average fuel usage to get the expected fuel usage for the remainder of
the lap.