Calculate fuel target
The fuel target is the current fuel level, minus the margin, divided by
the number of additional laps we want to complete in the stint.

To compute the number of additional laps we want to complete, we have to
know the lap that we pitted out on after refuelling. This changes the
LastPitLap synthetic field to accurately reflect the last out lap.